Fixed all warnings and some standards refactoring
Note, this commit adds graphical corruption to the sprite
batch, so I'm committing it to a branch.

- Enabled -Wall and treat warnings as errors
	- Most changes are using explicit casts
	- Added UNUSED macro to supress unused variable warnings
- Replaced 'static const' with 'static constexpr' where relevant
- Changed some 'enum' to 'enum class' with exlicit underlying types
- Added FLAG_OPS macro to define bitwise operators for an enum used to represent bitfield flags

+++ Momo/Include/Input/InputQueue.h
@@ -15,20 +15,20 @@ namespace Momo
 		public:
 			static Event kNullEvent;
 
-			InputQueue(int size = 256);
+			InputQueue(size_t size = 256);
 			~InputQueue();
 
 			void Push(Event& event);
 			const Event& Pop();
 
-			short Count() const { return mCount; }
+			size_t Count() const { return mCount; }
 
 		private:
-			int mSize;
+			size_t mSize;
 			Event* mQueue;
 
 			Event* mpStart;
-			int mCount;
+			size_t mCount;
 		};
 
 	}
